#179 `raise_if_term_is_qname` → `iolanta.parse_quads`
